refactor(brief_engine): extract Mode 3 primitives into aragora/brief_engine/
Phase 1 of the SecurityReportBrief rollout — lift the generic Protocol
B primitives (lifecycle, storage, budget, panel config, executor,
worker) out of aragora/pdb/ so future brief variants
(SecurityReportBrief, license audit, compliance review) can be built
without copy-paste.

See docs/plans/2026-04-22-security-report-brief-design.md §Phase 1.

Move/rename table:

  aragora/pdb/brief_state.py   → aragora/brief_engine/lifecycle.py
  aragora/pdb/storage.py       → aragora/brief_engine/storage.py
  aragora/pdb/budget.py        → aragora/brief_engine/budget.py
  aragora/pdb/panel_config.py  → aragora/brief_engine/panel_config.py
  aragora/pdb/protocol.py::run_protocol_b + contracts
                               → aragora/brief_engine/protocol.py
  aragora/pdb/worker.py        → aragora/brief_engine/worker.py

Within brief_engine, the PDB-namespaced types are renamed to generic
names: PDBExecutionInput → BriefExecutionInput, PDBPanelConfig →
BriefPanelConfig, PDBBudgetDecision → BriefBudgetDecision, etc. The
executor entry point is run_brief_protocol_b. The PDB package
re-exports all of these under the legacy PDB names and wraps
run_brief_protocol_b as run_protocol_b with the PDB prompt renderer +
default config injected — zero signature change, zero behaviour
change.

Stays PDB-specific (unchanged):
- aragora/pdb/prompts.py (PR-review prompt templates)
- aragora/pdb/input_loader.py (gh-backed PR input builder)
- aragora/config/pdb_panel.yaml (PDB panel config)

Tests:
- tests/pdb/ — 197 tests, all pass unchanged
- tests/server/handlers/test_review_queue_brief.py — 12 tests, all pass
- tests/brief_engine/test_pdb_compat.py — 25 new tests pinning the
  back-compat shim contract

Lint + format + mypy all clean on aragora/brief_engine/, aragora/pdb/,
and tests/brief_engine/.
